trying to figure out a solution to json dist issue...
"converting to JSON => sending over network => converting JSON to original types" comes with the issue that some data may not be converted back into correct type e.g. string "5" will be received as an int 5. due to complexity of JSONable data (e.g. dict within dict within dict...) makes a simple check of type not possible.

current idea:
- for distributed use, simple setup will force use of a JSONable Validator that ensures data can be serialized with JSON library before it enters the engine
- predicate classes changed to have: 1) "instance of" class for checking class type; and 2) "type cast" class for attempting to cast data to a type, ensuring that the cast will not throw an exception when casting in the predicate function

still need to test all of this... the tests in "tests" have yet to be fixed until I settle on a solution to this problem.